+++ src/projectgoldstarscalculators/Circles.java
@@ -7,7 +7,7 @@ import javax.swing.JOptionPane;
 import javax.swing.JTextField;
 public class Circles
 {
-    public static JTextField radiusField;
+    public static JTextField radiusField, areaField, circumferenceField;
     
     public Circles()
     {
@@ -17,14 +17,21 @@ public class Circles
     private void circles()
     {
         ProgramWindow frame = new ProgramWindow("Circles");
-        frame.setLayout(new GridLayout(2, 2));
-        frame.setSize(650 * ProjectGoldStarsCalculatorS.multiplier, 350 * ProjectGoldStarsCalculatorS.multiplier);
-        frame.setInstructionsMenuBar("Please enter the following information:");
+        frame.setLayout(new GridLayout(3, 3));
+        frame.setSize(800 * ProjectGoldStarsCalculatorS.multiplier, 425 * ProjectGoldStarsCalculatorS.multiplier);
+        frame.setInstructionsMenuBar("Please enter the radius of the circle, and then press Calculate to find its area and circumference:");
         frame.add(Components.standardLabel("Radius:"));
         setupRadiusField();
         frame.add(radiusField);
-        frame.add(new JLabel());
         frame.add(Components.button2("Calculate", new CalculateListener()));
+        frame.add(Components.standardLabel("Area:"));
+        setupAreaField();
+        frame.add(areaField);
+        frame.add(new JLabel());
+        frame.add(Components.standardLabel("Circumference:"));
+        setupCircumferenceField();
+        frame.add(circumferenceField);
+        frame.add(Components.disclaimerLabel("Note: Values may not be exact."));
         frame.makeVisible();
     }
     
@@ -33,6 +40,20 @@ public class Circles
         radiusField = new JTextField("0");
         radiusField.setFont(ProjectGoldStarsCalculatorS.bodyText2);
         radiusField.addActionListener(new CalculateListener());
+    }
+    
+    private void setupAreaField()
+    {
+        areaField = new JTextField("0");
+        areaField.setEditable(false);
+        areaField.setFont(ProjectGoldStarsCalculatorS.bodyText2);
+    }
+    
+    private void setupCircumferenceField()
+    {
+        circumferenceField = new JTextField("0");
+        circumferenceField.setEditable(false);
+        circumferenceField.setFont(ProjectGoldStarsCalculatorS.bodyText2);
     }
     
     public static class CalculateListener implements ActionListener
@@ -66,9 +87,8 @@ public class Circles
             //Calculate the circumference.
             double circumference = 2 * Math.PI * radiusD;
             //Display the area and circumference.
-            output = "Area of the Circle: " + area + "\n"
-                    + "Circumference of the Circle: " + circumference;
-            JOptionPane.showMessageDialog(null, output, "Circles", JOptionPane.INFORMATION_MESSAGE);
+            areaField.setText("" + area);
+            circumferenceField.setText("" + circumference);
         }
     }
 }